backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-allauth==0.62.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-decouple==3.8
django-storages==1.14.2
boto3==1.34.69
Pillow==10.3.0
stripe==9.0.0
django-anymail[brevo]==10.3
django-filter==24.2
django-extensions==3.2.3
django-apscheduler==0.6.2
anthropic
yookassa==3.3.0
gunicorn==21.2.0
whitenoise==6.6.0
django-jazzmin==3.0.1